fix(objects): Refactor Brep classes into it's own files
