+++ Objects/Objects/Geometry/Brep.cs
@@ -509,141 +509,6 @@ namespace Objects.Geometry
     Unknown = 2
   }
 
-  /// <summary>
-  /// Represents a face on a <see cref="Brep"/>
-  /// </summary>
-  public class BrepFace : Base
-  {
-    [JsonIgnore] public Brep Brep { get; set; }
-    public int SurfaceIndex { get; set; }
-    public List<int> LoopIndices { get; set; }
-    public int OuterLoopIndex { get; set; }
-    public bool OrientationReversed { get; set; }
-
-    public BrepFace()
-    {
-    }
-
-    public BrepFace(Brep brep, int surfaceIndex, List<int> loopIndices, int outerLoopIndex, bool orientationReversed)
-    {
-      Brep = brep;
-      SurfaceIndex = surfaceIndex;
-      LoopIndices = loopIndices;
-      OuterLoopIndex = outerLoopIndex;
-      OrientationReversed = orientationReversed;
-    }
-
-    [JsonIgnore] public BrepLoop OuterLoop => Brep.Loops[OuterLoopIndex];
-    [JsonIgnore] public Surface Surface => Brep.Surfaces[SurfaceIndex];
-    [JsonIgnore] public List<BrepLoop> Loops => LoopIndices.Select(i => Brep.Loops[i]).ToList();
-  }
-
-  /// <summary>
-  /// Represents a UV Trim Closed Loop on one of the <see cref="Brep"/>'s surfaces.
-  /// </summary>
-  public class BrepLoop : Base
-  {
-    [JsonIgnore] public Brep Brep { get; set; }
-    public int FaceIndex { get; set; }
-    public List<int> TrimIndices { get; set; }
-    public BrepLoopType Type { get; set; }
-
-    public BrepLoop()
-    {
-    }
-
-    public BrepLoop(Brep brep, int faceIndex, List<int> trimIndices, BrepLoopType type)
-    {
-      Brep = brep;
-      FaceIndex = faceIndex;
-      TrimIndices = trimIndices;
-      Type = type;
-    }
-
-    [JsonIgnore] public BrepFace Face => Brep.Faces[FaceIndex];
-    [JsonIgnore] public List<BrepTrim> Trims => TrimIndices.Select(i => Brep.Trims[i]).ToList();
-  }
-
-  /// <summary>
-  /// Represents a UV Trim curve for one of the <see cref="Brep"/>'s surfaces.
-  /// </summary>
-  public class BrepTrim : Base
-  {
-    [JsonIgnore] public Brep Brep { get; set; }
-    public int EdgeIndex { get; set; }
-    public int StartIndex { get; set; }
-    public int EndIndex { get; set; }
-    public int FaceIndex { get; set; }
-    public int LoopIndex { get; set; }
-    public int CurveIndex { get; set; }
-    public int IsoStatus { get; set; }
-    public BrepTrimType TrimType { get; set; }
-    public bool IsReversed { get; set; }
-
-    public Interval Domain { get; set; }
-
-    public BrepTrim()
-    {
-    }
-
-    public BrepTrim(Brep brep, int edgeIndex, int faceIndex, int loopIndex, int curveIndex, int isoStatus,
-      BrepTrimType trimType, bool reversed, int startIndex, int endIndex)
-    {
-      Brep = brep;
-      EdgeIndex = edgeIndex;
-      FaceIndex = faceIndex;
-      LoopIndex = loopIndex;
-      CurveIndex = curveIndex;
-      IsoStatus = isoStatus;
-      TrimType = trimType;
-      IsReversed = reversed;
-    }
-
-    [JsonIgnore] public BrepFace Face => Brep.Faces[FaceIndex];
-
-    [JsonIgnore] public BrepLoop Loop => Brep.Loops[LoopIndex];
-
-    [JsonIgnore] public BrepEdge Edge => EdgeIndex != -1 ? Brep.Edges[EdgeIndex] : null;
-
-    [JsonIgnore] public ICurve Curve2d => Brep.Curve2D[CurveIndex];
-  }
-
-  /// <summary>
-  /// Represents an edge of the <see cref="Brep"/>.
-  /// </summary>
-  public class BrepEdge : Base
-  {
-    [JsonIgnore] public Brep Brep { get; set; }
-    public int Curve3dIndex { get; set; }
-    public int[] TrimIndices { get; set; }
-    public int StartIndex { get; set; }
-    public int EndIndex { get; set; }
-
-    public bool ProxyCurveIsReversed { get; set; }
-
-    public Interval Domain { get; set; }
-    public BrepEdge()
-    {
-    }
-
-    public BrepEdge(Brep brep, int curve3dIndex, int[] trimIndices, int startIndex, int endIndex,
-      bool proxyCurvedIsReversed, Interval domain)
-    {
-      Brep = brep;
-      Curve3dIndex = curve3dIndex;
-      TrimIndices = trimIndices;
-      StartIndex = startIndex;
-      EndIndex = endIndex;
-      ProxyCurveIsReversed = proxyCurvedIsReversed;
-      Domain = domain;
-    }
-
-    [JsonIgnore] public Point StartVertex => Brep.Vertices[StartIndex];
-    [JsonIgnore] public Point EndVertex => Brep.Vertices[EndIndex];
-    [JsonIgnore] public IEnumerable<BrepTrim> Trims => TrimIndices.Select(i => Brep.Trims[i]);
-    [JsonIgnore] public ICurve Curve => Brep.Curve3D[Curve3dIndex];
-  }
-
 
   /// <summary>
   /// Represents the type of a loop in a <see cref="Brep"/>'s face.
